requirements.txt
google_api_python_client==2.73.0
httplib2==0.21.0
numpy==1.23.4
oauth2client==4.1.3
openpyxl==3.0.10
pandas==1.5.1
Pillow==9.4.0
psycopg2==2.9.5
requests==2.28.1
toloka_kit==0.1.23
tqdm==4.64.1
